lavc/vvc: Assert coding tree root is within bounds
If this assumption is violated, which I have seen happen in a couple of
cases now from fuzzed bitstreams, there will inevitably be a crash.  In
the best case, as CTUs can't be split to get within bounds, there will
be an infinite loop and stack overflow.  In the worst case, in can lead
to a segmentation fault.