lavu/common: Fix AV_CEIL_RSHIFT for unsigned LHS
The first branch of this ternary expression was intended to avoid
having two shift operations in the case the RHS is not known at
compile time.  It only works if the LHS has a signed type however,
otherwise the result is invalid.

We could alternatively have different versions of AV_CEIL_RSHIFT for
different sizes of operand, then cast the LHS to the relevant signed
type.